Add files via upload
This is a JSP Login Authentication Web Application.It contains a Login, Registration and Admin and Welcome page. The login requires a username (email) and password. New Users may Register and will then be redirected to login. The Admin page is free to access and containss a table of all users and their login info. This info except id may be edited or deleted. The welcome page is created with the logged in users session and allows user to logout.